User_Services/user_producer/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Muhammad Mateen Mirza <teenmeem@google.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
psycopg = {extras = ["binary"], version = "^3.2.1"}
sqlmodel = "^0.0.19"
logging = "^0.4.9.6"
aiokafka = "^0.10.0"
kafka-utils = "^4.1.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = "^3.3.0"

[tool.poetry.group.dev.dependencies]
mypy = "^1.10.0"
pytest = "^8.2.1"
types-passlib = "^1.7.7.20240327"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
